bugfix for moon rise/set based on timezone
Moon rise/set were incorrect on some dates - it was still shoring a time even if the moon didn't rise or set that day.

This was due to incorrect documentation for suncalc.js (it references an .alwaysUp property when it really just doesn't include a .set property when it doesn't set).  I fixed this.

The second issue is inconsistency handling timezones - suncalc uses local time to determine the moon rise/set for a particular date, but you might be trying to look at a calendar for a different timezone.  It's better now, but still not perfect.